line-bot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
line-bot-sdk==3.14.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
anthropic==0.44.0
redis>=5.0.0
Pillow==11.1.0